+++ src/constants/button.ts
@@ -1,5 +1,11 @@
 import { SelectableValue } from '@grafana/data';
-import { CapitalizeFirstLetter } from '../constant-utils';
+
+/**
+ * Capitalize First Letter
+ */
+export const CapitalizeFirstLetter = (str: string) => {
+  return str.charAt(0).toUpperCase() + str.slice(1);
+};
 
 /**
  * Button Variants