+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Update Form Elements to delay save changes with auto-save (#186)
 - Migrate to Plugin Tools 1.5.0 (#187)
 - Update to Node 18 (#188)
+- Update constants and E2E tests (#190)
 
 ## 2.8.0 (2023-03-16)